Add files via upload
Our first commit is the code to complete to tasks. One of these tasks are knocking the down the stack of 3 processed iron boxes and the 3 processed quicklime boxes. The second task is making the robot knock out the mineral transport from under the mining ramp. However, this code is still not finished. The remaining tasks are: to move the mineral transport under the platform where the water is (royal blue pom poms) and the second task is lifting the claw to knock the water (royal blue pom poms) into the mineral transport.