Add files via upload
This commit is the same code as our first commit, but with minor changes. The first change is on line 44 and the second is on line 49. The change on line 44 is the amount of msleep, instead of being 1500, we changed it to 2700. The second change on line 49 is also msleep, instead of being 1500, we changed it to 2300.